feat: 고철 가격 조회 추가

+++ src/main/java/com/kdt_proj2_be/domain/ScrapPrice.java
@@ -17,7 +17,7 @@ public class ScrapPrice {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer scrapPriceId;
+    private Long scrapPriceId;
 
     @ManyToOne
     @JoinColumn(name = "scrap_type_id")
